Reorganized code & improved logging
